dev-requirements.txt
setuptools>=64
setuptools_scm>=8
build ==1.* # conda: python-build
ruff >= 0.2
copier ==9.*
unbeheader ~= 1.3 # conda: @pip
ipython >= 7
pyarrow>=15
sphinx-autobuild >= 2021
docopt >= 0.6
tomlkit>=0.12
pyyaml==6.*
packaging>=24.0
tqdm >= 4
pandas-stubs >=1.5,<3
-r test-requirements.txt
-r doc-requirements.txt
-r demo-requirements.txt
